move notebook page concerns into page, app bits keep app bits
